backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
python-jose[cryptography]==3.3.0
pydantic==2.9.0
python-multipart==0.0.9
httpx==0.27.0
keybert==0.8.5
sentence-transformers==3.0.0
transformers==4.44.0
torch==2.4.0
Pillow==10.4.0
numpy==1.26.4